从被监视端直接转成JSON格式，CMD_Client_Dump_InvokeMethod、CMD_Server_Dump_InvokeMethod

+++ como/runtime/rpc/ZeroMQ/ThreadPoolZmqActor.cpp
@@ -319,12 +319,12 @@ HandleMessage_Object_ReleasePeer:
                 }
 
                 switch (rtmCommand->command) {
-                    case (Short)RTM_CommandType::CMD_Server_Activate_InvokeMethod: {
+                    case RTM_CommandType::CMD_Server_Activate_InvokeMethod: {
                         CStub::From(worker->mStub)->mMonitorInvokeMethod = true;
                         goto HandleMessage_RuntimeMonitor;
                     }
 
-                    case (Short)RTM_CommandType::CMD_Server_Deactivate_InvokeMethod: {
+                    case RTM_CommandType::CMD_Server_Deactivate_InvokeMethod: {
                         CStub::From(worker->mStub)->mMonitorInvokeMethod = false;
                         goto HandleMessage_RuntimeMonitor;
                     }
